LibRuntime: Move LibC initialization routines from LibC
